Rag/requirements.txt
streamlit
langchain
langchain-community
langchain-cohere
langchain-chroma
python-dotenv
pymupdf
chromadb